scraper/requirements.txt
requests
beautifulsoup4
pymongo
python-dotenv

